New image processing flow in StateProcessor: serialize all stuff -> pass to separate job (for now in the same process) -> deserialize all stuff -> do processing
