+++ lib/browser-pool/basic-pool.js
@@ -23,7 +23,7 @@ util.inherits(BasicPool, Pool);
 
 Pool.prototype.getBrowser = function(id) {
     var _this = this,
-        browser = new Browser(this._config.forBrowser(id)),
+        browser = Browser.create(this._config.forBrowser(id)),
         launchPromise = browser.launch(this._calibrator);
 
     activeSessions[browser.id] = {
