+++ lib/browser-pool/limited-use-set.js
@@ -1,6 +1,5 @@
 'use strict';
-var WeakMap = require('es6-weak-map'),
-    q = require('q'),
+var q = require('q'),
     log = require('debug')('gemini:limited-use-set'),
     _ = require('lodash'),
     MetaError = require('../errors/meta-error');
