+++ lib/capture-session.js
@@ -8,9 +8,10 @@ var util = require('util'),
     promiseUtil = require('q-promise-utils'),
     ActionsBuilder = require('./tests-api/actions-builder'),
     StateError = require('./errors/state-error'),
+    Browser = require('./browser'),
     temp = require('./temp');
 
-module.exports = inherit({
+var CaptureSession = inherit({
     __constructor: function(browser) {
         this.browser = browser;
         this.log = debug('gemini:capture:' + this.browser.id);
@@ -54,6 +55,12 @@ module.exports = inherit({
             })
             .fail(_.noop)
             .thenResolve(obj);
+    },
+
+    serialize: function() {
+        return {
+            browser: this.browser.serialize()
+        };
     },
 
     _cropImage: function(screenImage, pageDisposition) {
@@ -172,7 +179,16 @@ module.exports = inherit({
             ? dimension * pixelRatio
             : dimension;
     }
+}, {
+    fromObject: function(serializedSession) {
+        return Browser.fromObject(serializedSession.browser)
+            .then(function(browser) {
+                return new CaptureSession(browser);
+            });
+    }
 });
+
+module.exports = CaptureSession;
 
 function isOutsideOfImage(imageSize, cropArea) {
     return cropArea.top < 0 || cropArea.left < 0 || cropArea.left + cropArea.width > imageSize.width;
