+++ lib/constants/wd-errors.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+    ELEMENT_NOT_FOUND: 7
+};
