+++ lib/image.js
@@ -68,7 +68,7 @@ var Image = inherit({
             diffOptions.tolerance = opts.tolerance;
         }
         return q.nfcall(looksSame.createDiff, diffOptions);
-    },
+    }
 });
 
 module.exports = Image;
