+++ lib/runner/index.js
@@ -94,7 +94,7 @@ var TestsRunner = inherit(Runner, {
         }
 
         if (!this._failCollector.tryToSubmitStateResult(result)) {
-            this.emit(this._stateProcessor.getProcessedCaptureEventName(), result);
+            this.emit(this._stateProcessor.jobDoneEvent, result);
         }
     },
