+++ lib/state-processor/capture-processor/capture-processor.js
@@ -1,13 +1,7 @@
 'use strict';
 
-var inherit = require('inherit');
-
-module.exports = inherit({
-    getProcessedCaptureEventName: function() {
-        throw new Error('Not implemented');
-    },
-
-    exec: function(capture, opts) {
+module.exports = class CaptureProcessor {
+    exec(capture, opts) {
         throw new Error('Not implemented');
     }
-});
+};
