+++ lib/state-processor/capture-processor/screen-shooter.js
@@ -2,17 +2,14 @@
 
 var fs = require('q-io/fs'),
     path = require('path'),
-    inherit = require('inherit'),
-    CaptureProcessor = require('./capture-processor'),
+    CaptureProcessor = require('./capture-processor');
 
-    RunnerEvents = require('../../constants/runner-events');
+module.exports = class ScreenShooter extends CaptureProcessor {
+    static create() {
+        return new ScreenShooter();
+    }
 
-module.exports = inherit(CaptureProcessor, {
-    getProcessedCaptureEventName: function() {
-        return RunnerEvents.CAPTURE;
-    },
-
-    exec: function(capture, opts) {
+    exec(capture, opts) {
         var savePath = opts.refPath;
 
         //TODO: create directories only once
@@ -22,4 +19,4 @@ module.exports = inherit(CaptureProcessor, {
             })
             .thenResolve({imagePath: savePath});
     }
-});
+};
