+++ lib/state-processor/capture-processor/screen-updater/diff-screen-updater.js
@@ -1,14 +1,13 @@
 'use strict';
 
 var fs = require('q-io/fs'),
-    inherit = require('inherit'),
     temp = require('../../../temp'),
 
     Image = require('../../../image'),
     ScreenUpdater = require('./screen-updater');
 
-module.exports = inherit(ScreenUpdater, {
-    _processCapture: function(capture, opts, isRefExists) {
+module.exports = class DiffScreenUpdater extends ScreenUpdater {
+    _processCapture(capture, opts, isRefExists) {
         if (!isRefExists) {
             return;
         }
@@ -30,4 +29,4 @@ module.exports = inherit(ScreenUpdater, {
                     });
             }.bind(this));
     }
-});
+};
