+++ lib/state-processor/capture-processor/screen-updater/meta-screen-updater.js
@@ -1,22 +1,23 @@
 'use strict';
 
-var inherit = require('inherit'),
-    ScreenUpdater = require('./screen-updater'),
+var ScreenUpdater = require('./screen-updater'),
     DiffUpdater = require('./diff-screen-updater'),
     NewUpdater = require('./new-screen-updater'),
 
     util = require('q-promise-utils');
 
-module.exports = inherit(ScreenUpdater, {
-    __constructor: function() {
+module.exports = class MetaScreenUpdater extends ScreenUpdater {
+    constructor() {
+        super();
+
         this._diffUpdater = new DiffUpdater();
         this._newUpdater = new NewUpdater();
-    },
+    }
 
-    _processCapture: function(capture, opts, isRefExists) {
+    _processCapture(capture, opts, isRefExists) {
         return util.sequence([
             this._diffUpdater._processCapture.bind(this._diffUpdater, capture, opts, isRefExists),
             this._newUpdater._processCapture.bind(this._newUpdater, capture, opts, isRefExists)
         ]);
     }
-});
+};
