+++ lib/state-processor/capture-processor/screen-updater/new-screen-updater.js
@@ -2,11 +2,10 @@
 
 var fs = require('q-io/fs'),
     path = require('path'),
-    inherit = require('inherit'),
     ScreenUpdater = require('./screen-updater');
 
-module.exports = inherit(ScreenUpdater, {
-    _processCapture: function(capture, opts, isRefExists) {
+module.exports = class NewScreenUpdater extends ScreenUpdater {
+    _processCapture(capture, opts, isRefExists) {
         if (isRefExists) {
             return;
         }
@@ -16,4 +15,4 @@ module.exports = inherit(ScreenUpdater, {
                 return capture.image.save(opts.refPath);
             });
     }
-});
+};
