+++ lib/state-processor/capture-processor/screen-updater/screen-updater.js
@@ -1,17 +1,10 @@
 'use strict';
 
-var inherit = require('inherit'),
-    fs = require('q-io/fs'),
-    CaptureProcessor = require('../capture-processor'),
+var fs = require('q-io/fs'),
+    CaptureProcessor = require('../capture-processor');
 
-    RunnerEvents = require('../../../constants/runner-events');
-
-module.exports = inherit(CaptureProcessor, {
-    getProcessedCaptureEventName: function() {
-        return RunnerEvents.CAPTURE;
-    },
-
-    exec: function(capture, opts) {
+module.exports = class ScreenUpdater extends CaptureProcessor {
+    exec(capture, opts) {
         var refPath = opts.refPath,
             _this = this;
 
@@ -20,9 +13,9 @@ module.exports = inherit(CaptureProcessor, {
                 return _this._processCapture(capture, opts, isRefExists);
             })
             .thenResolve({imagePath: refPath});
-    },
+    }
 
-    _processCapture: function() {
+    _processCapture() {
         throw new Error('Not implemented');
     }
-});
+};
