+++ lib/state-processor/capture-processor/tester.js
@@ -1,26 +1,25 @@
 'use strict';
 
 var q = require('q'),
-    inherit = require('inherit'),
     fs = require('q-io/fs'),
     temp = require('../../temp'),
 
     CaptureProcessor = require('./capture-processor'),
     Image = require('../../image'),
-    NoRefImageError = require('../../errors/no-ref-image-error'),
+    NoRefImageError = require('../../errors/no-ref-image-error');
 
-    RunnerEvents = require('../../constants/runner-events');
+module.exports = class Tester extends CaptureProcessor {
+    static create(diffColor) {
+        return new Tester(diffColor);
+    }
 
-module.exports = inherit(CaptureProcessor, {
-    __constructor: function(diffColor) {
+    constructor(diffColor) {
+        super();
+
         this._diffColor = diffColor;
-    },
+    }
 
-    getProcessedCaptureEventName: function() {
-        return RunnerEvents.END_TEST;
-    },
-
-    exec: function(capture, opts) {
+    exec(capture, opts) {
         var refPath = opts.refPath,
             tmpPath = temp.path({suffix: '.png'}),
             tolerance = opts.tolerance,
@@ -58,4 +57,4 @@ module.exports = inherit(CaptureProcessor, {
                 };
             });
     }
-});
+};
