+++ lib/state-processor/index.js
@@ -1,20 +1,25 @@
 'use strict';
 
-var Tester = require('./capture-processor/tester'),
-    ScreenShooter = require('./capture-processor/screen-shooter'),
-    ScreenUpdater = require('./capture-processor/screen-updater'),
-    StateProcessor = require('./state-processor');
+var StateProcessor = require('./state-processor'),
+    RunnerEvents = require('../constants/runner-events');
+
+function create(captureProcessorName, jobDoneEvent, constructorArgs) {
+    return new StateProcessor({
+            module: require.resolve('./capture-processor/' + captureProcessorName),
+            constructorArgs
+        }, jobDoneEvent);
+}
 
 module.exports = {
     createTester: function(config) {
-        return new StateProcessor(new Tester(config.system.diffColor));
+        return create('tester', RunnerEvents.END_TEST, config.system.diffColor);
     },
 
     createScreenShooter: function() {
-        return new StateProcessor(new ScreenShooter());
+        return create('screen-shooter', RunnerEvents.CAPTURE);
     },
 
     createScreenUpdater: function(options) {
-        return new StateProcessor(ScreenUpdater.create(options));
+        return create('screen-updater', RunnerEvents.CAPTURE, options);
     }
 };
