+++ lib/state-processor/state-processor.js
@@ -2,15 +2,13 @@
 
 var inherit = require('inherit'),
     _ = require('lodash'),
-    q = require('q');
+    q = require('q'),
+    job = require('./job');
 
 module.exports = inherit({
-    __constructor: function(captureProcessor) {
-        this._captureProcessor = captureProcessor;
-    },
-
-    getProcessedCaptureEventName: function() {
-        return this._captureProcessor.getProcessedCaptureEventName();
+    __constructor: function(captureProcessorInfo, jobDoneEvent) {
+        this._captureProcessorInfo = captureProcessorInfo;
+        this.jobDoneEvent = jobDoneEvent;
     },
 
     prepare: function(emitter) {
@@ -18,21 +16,18 @@ module.exports = inherit({
     },
 
     exec: function(state, browserSession, pageDisposition) {
-        var execOpts = {
-            refPath: browserSession.browser.config.getScreenshotPath(state.suite, state.name),
-            tolerance: _.isNumber(state.tolerance)
-                ? state.tolerance
-                : browserSession.browser.config.tolerance
-        };
+        var jobArgs = {
+                captureProcessorInfo: this._captureProcessorInfo,
+                browserSession: browserSession.serialize(),
+                pageDisposition: pageDisposition,
+                execOpts: {
+                    refPath: browserSession.browser.config.getScreenshotPath(state.suite, state.name),
+                    tolerance: _.isNumber(state.tolerance)
+                        ? state.tolerance
+                        : browserSession.browser.config.tolerance
+                }
+            };
 
-        return browserSession.capture(pageDisposition)
-            .then(function(capture) {
-                return this._captureProcessor.exec(capture, execOpts)
-                    .then(function(processed) {
-                        return _.extend(processed, {
-                            coverage: capture.coverage
-                        });
-                    });
-            }.bind(this));
+        return job(jobArgs);
     }
 });
