+++ lib/suite-monitor.js
@@ -3,7 +3,6 @@ var EventEmitter = require('events').EventEmitter,
 
     _ = require('lodash'),
     inherit = require('inherit'),
-    WeakMap = require('es6-weak-map'),
 
     RunnerEvents = require('./constants/runner-events');
